+++ src/main/kotlin/org/jetbrains/intellij/tasks/DownloadAndroidStudioProductReleasesXmlTask.kt
@@ -2,11 +2,12 @@
 
 package org.jetbrains.intellij.tasks
 
+import org.gradle.api.provider.Property
 import org.gradle.api.tasks.CacheableTask
+import org.gradle.api.tasks.Input
+import org.gradle.api.tasks.Optional
 import org.gradle.api.tasks.Sync
 import org.jetbrains.intellij.IntelliJPluginConstants
-import org.jetbrains.intellij.IntelliJPluginConstants.ANDROID_STUDIO_PRODUCTS_RELEASES_URL
-import org.jetbrains.intellij.error
 import org.jetbrains.intellij.logCategory
 
 @CacheableTask
@@ -14,20 +15,16 @@ abstract class DownloadAndroidStudioProductReleasesXmlTask : Sync() {
 
     private val context = logCategory()
 
+    /**
+     * Represents the property that holds the URL for releases.
+     * Default value: [IntelliJPluginConstants.ANDROID_STUDIO_PRODUCTS_RELEASES_URL]
+     */
+    @get:Input
+    @get:Optional
+    abstract val releasesUrl: Property<String>
+
     init {
         group = IntelliJPluginConstants.PLUGIN_GROUP_NAME
         description = "Downloads XML files containing the Android Studio product release information."
-
-        // TODO: migrate to `project.resources.binary` whenever it's available. Ref: https://github.com/gradle/gradle/issues/25237
-        from(
-            project.resources.text
-                .fromUri(ANDROID_STUDIO_PRODUCTS_RELEASES_URL)
-                .runCatching { asFile("UTF-8") }
-                .onFailure { error(context, "Cannot resolve product releases", it) }
-                .getOrDefault("<products />")
-        ) {
-            rename { "android_studio_product_releases.xml" }
-        }
-        into(temporaryDir)
     }
 }